chore: hoist lib to root
