+++ app/api/route.ts
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server";
 import { calculationSchema } from "../schemas/calculationSchema";
 import * as calculationService from "../services/calculationService";
 import calculateFairhold from "../models/calculateFairhold";
-import { APIError } from "../lib/exceptions";
+import { APIError } from "../../lib/calculator/exceptions";
 
 export async function POST(req: Request) {
   try {
