+++ components/custom/dashboard/graphs/CostOverTimeStackedBarChart.tsx
@@ -4,11 +4,11 @@ import { Bar, BarChart, CartesianGrid, XAxis, YAxis, Tooltip } from "recharts";
 import { Card, CardContent } from "@/components/card";
 import {
   StyledChartContainer,
-} from "../ui/StyledChartContainer";
+} from "@/components/custom/ui/StyledChartContainer";
 import { TooltipProps } from "recharts";
 import { ValueType } from "tailwindcss/types/config";
 import { NameType } from "recharts/types/component/DefaultTooltipContent";
-import { formatValue } from "@/app/lib/format";
+import { formatValue } from "@/lib/calculator/format";
 
 export interface LifetimeBarData {
   landRent?: number;
