+++ lib/survey/types.tsx
